requirements.txt
# Flask Backend
Flask==3.0.0
flask-cors==4.0.0

# Core Dependencies
spacy==3.7.2
scikit-learn
google-generativeai
python-dotenv>=0.19.0
pandas==2.1.4
numpy==1.26.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0

# Model Persistence
joblib==1.3.2

# Additional Utilities
nltk==3.8.1
